+++ src/dolphin/workflows/displacement.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 from __future__ import annotations
 
+import contextlib
 import multiprocessing as mp
 from collections import defaultdict
 from concurrent.futures import ProcessPoolExecutor
@@ -10,24 +11,16 @@ from pathlib import Path
 from pprint import pformat
 from typing import Mapping, NamedTuple, Sequence
 
-from opera_utils import group_by_burst, group_by_date
+from opera_utils import get_dates, group_by_burst, group_by_date
 from tqdm.auto import tqdm
 
-from dolphin import __version__
+from dolphin import __version__, io, timeseries, utils
 from dolphin._log import get_log, log_runtime
 from dolphin.atmosphere import estimate_ionospheric_delay, estimate_tropospheric_delay
-from dolphin.io import get_raster_bounds, get_raster_crs
-from dolphin.utils import (
-    DummyProcessPoolExecutor,
-    disable_gpu,
-    get_max_memory_usage,
-    prepare_geometry,
-    set_num_threads,
-)
 
 from . import stitching_bursts, unwrapping, wrapped_phase
 from ._utils import _create_burst_cfg, _remove_dir_if_empty
-from .config import DisplacementWorkflow
+from .config import DisplacementWorkflow, TimeseriesOptions
 
 logger = get_log(__name__)
 
@@ -40,6 +33,7 @@ class OutputPaths(NamedTuple):
     stitched_cor_paths: list[Path]
     stitched_temp_coh_file: Path
     stitched_ps_file: Path
+    stitched_amp_dispersion_file: Path
     unwrapped_paths: list[Path] | None
     conncomp_paths: list[Path] | None
     tropospheric_corrections: list[Path] | None
@@ -67,8 +61,8 @@ def run(
     logger.debug(pformat(cfg.model_dump()))
 
     if not cfg.worker_settings.gpu_enabled:
-        disable_gpu()
-    set_num_threads(cfg.worker_settings.threads_per_worker)
+        utils.disable_gpu()
+    utils.set_num_threads(cfg.worker_settings.threads_per_worker)
 
     try:
         grouped_slc_files = group_by_burst(cfg.cslc_file_list)
@@ -140,6 +134,7 @@ def run(
     ifg_file_list: list[Path] = []
     temp_coh_file_list: list[Path] = []
     ps_file_list: list[Path] = []
+    amp_dispersion_file_list: list[Path] = []
     # The comp_slc tracking object is a dict, since we'll need to organize
     # multiple comp slcs by burst (they'll have the same filename)
     comp_slc_dict: dict[str, list[Path]] = {}
@@ -148,7 +143,7 @@ def run(
     Executor = (
         ProcessPoolExecutor
         if cfg.worker_settings.n_parallel_bursts > 1
-        else DummyProcessPoolExecutor
+        else utils.DummyProcessPoolExecutor
     )
     mw = cfg.worker_settings.n_parallel_bursts
     ctx = mp.get_context("spawn")
@@ -173,11 +168,12 @@ def run(
         for fut in fut_to_burst:
             burst = fut_to_burst[fut]
 
-            cur_ifg_list, comp_slcs, temp_coh, ps_file = fut.result()
+            cur_ifg_list, comp_slcs, temp_coh, ps_file, amp_disp_file = fut.result()
             ifg_file_list.extend(cur_ifg_list)
             comp_slc_dict[burst] = comp_slcs
             temp_coh_file_list.append(temp_coh)
             ps_file_list.append(ps_file)
+            amp_dispersion_file_list.append(amp_disp_file)
 
     # ###################################
     # 2. Stitch burst-wise interferograms
@@ -192,10 +188,12 @@ def run(
         stitched_cor_paths,
         stitched_temp_coh_file,
         stitched_ps_file,
+        stitched_amp_dispersion_file,
     ) = stitching_bursts.run(
         ifg_file_list=ifg_file_list,
         temp_coh_file_list=temp_coh_file_list,
         ps_file_list=ps_file_list,
+        amp_dispersion_list=amp_dispersion_file_list,
         stitched_ifg_dir=cfg.interferogram_network._directory,
         output_options=cfg.output_options,
         file_date_fmt=cfg.input_options.cslc_date_fmt,
@@ -214,6 +212,7 @@ def run(
             stitched_cor_paths=stitched_cor_paths,
             stitched_temp_coh_file=stitched_temp_coh_file,
             stitched_ps_file=stitched_ps_file,
+            stitched_amp_dispersion_file=stitched_amp_dispersion_file,
             unwrapped_paths=None,
             conncomp_paths=None,
             tropospheric_corrections=None,
@@ -231,7 +230,25 @@ def run(
     )
 
     # ##############################################
-    # 4. Estimate corrections for each interferogram
+    # 4. If a network was unwrapped, invert it
+    # ##############################################
+
+    ts_opts = cfg.timeseries_options
+    if ts_opts.run_inversion or ts_opts.run_velocity:
+        inverted_phase_paths = run_timeseries(
+            ts_opts=ts_opts,
+            unwrapped_paths=unwrapped_paths,
+            conncomp_paths=conncomp_paths,
+            cor_paths=stitched_cor_paths,
+            stitched_amp_dispersion_file=stitched_amp_dispersion_file,
+            # TODO: do i care to configure block shape, or num threads from somewhere?
+            # num_threads=cfg.worker_settings....?
+        )
+    else:
+        inverted_phase_paths = unwrapped_paths
+
+    # ##############################################
+    # 5. Estimate corrections for each interferogram
     # ##############################################
     tropo_paths: list[Path] | None = None
     iono_paths: list[Path] | None = None
@@ -245,10 +262,10 @@ def run(
         # Prepare frame geometry files
         geometry_dir = out_dir / "geometry"
         geometry_dir.mkdir(exist_ok=True)
-        crs = get_raster_crs(ifg_filenames[0])
+        crs = io.get_raster_crs(ifg_filenames[0])
         epsg = crs.to_epsg()
-        out_bounds = get_raster_bounds(ifg_filenames[0])
-        frame_geometry_files = prepare_geometry(
+        out_bounds = io.get_raster_bounds(ifg_filenames[0])
+        frame_geometry_files = utils.prepare_geometry(
             geometry_dir=geometry_dir,
             geo_files=cfg.correction_options.geometry_files,
             matching_file=ifg_filenames[0],
@@ -302,7 +319,9 @@ def run(
         stitched_cor_paths=stitched_cor_paths,
         stitched_temp_coh_file=stitched_temp_coh_file,
         stitched_ps_file=stitched_ps_file,
-        unwrapped_paths=unwrapped_paths,
+        stitched_amp_dispersion_file=stitched_amp_dispersion_file,
+        # unwrapped_paths=unwrapped_paths,
+        unwrapped_paths=inverted_phase_paths,
         conncomp_paths=conncomp_paths,
         tropospheric_corrections=tropo_paths,
         ionospheric_corrections=iono_paths,
@@ -311,7 +330,82 @@ def run(
 
 def _print_summary(cfg):
     """Print the maximum memory usage and version info."""
-    max_mem = get_max_memory_usage(units="GB")
+    max_mem = utils.get_max_memory_usage(units="GB")
     logger.info(f"Maximum memory usage: {max_mem:.2f} GB")
     logger.info(f"Config file dolphin version: {cfg._dolphin_version}")
     logger.info(f"Current running dolphin version: {__version__}")
+
+
+def run_timeseries(
+    ts_opts: TimeseriesOptions,
+    unwrapped_paths: Sequence[Path],
+    conncomp_paths: Sequence[Path],
+    cor_paths: Sequence[Path],
+    stitched_amp_dispersion_file: Path,
+    num_threads: int = 5,
+) -> list[Path]:
+    """Invert the unwrapped interferograms, estimate timeseries and phase velocity."""
+    output_path = ts_opts._directory
+    output_path.mkdir(exist_ok=True, parents=True)
+
+    # First we find the reference point for the unwrapped interferograms
+    reference = timeseries.select_reference_point(
+        conncomp_paths,
+        stitched_amp_dispersion_file,
+        output_dir=output_path,
+    )
+
+    ifg_date_pairs = [get_dates(f) for f in unwrapped_paths]
+    sar_dates = sorted(set(utils.flatten(ifg_date_pairs)))
+    # if we did single-reference interferograms, for `n` sar dates, we will only have
+    # `n-1` interferograms. Any more than n-1 ifgs means we need to invert
+    needs_inversion = len(unwrapped_paths) > len(sar_dates) - 1
+    # check if we even need to invert, or if it was single reference
+    inverted_phase_paths: list[Path] = []
+    if needs_inversion:
+        logger.info("Selecting a reference point for unwrapped interferograms")
+
+        logger.info("Inverting network of %s unwrapped ifgs", len(unwrapped_paths))
+        inverted_phase_paths = timeseries.invert_unw_network(
+            unw_file_list=unwrapped_paths,
+            reference=reference,
+            output_dir=output_path,
+        )
+    else:
+        logger.info(
+            "Skipping inversion step: only single reference interferograms exist."
+        )
+        # Symlink the unwrapped paths to `timeseries/`
+        for p in unwrapped_paths:
+            target = output_path / p.name
+            with contextlib.suppress(FileExistsError):
+                target.symlink_to(p)
+            inverted_phase_paths.append(target)
+        # Make extra "0" raster so that the number of rasters matches len(sar_dates)
+        ref_raster = output_path / (
+            utils.format_dates(sar_dates[0], sar_dates[0]) + ".tif"
+        )
+        io.write_arr(
+            arr=None, output_name=ref_raster, like_filename=inverted_phase_paths[0]
+        )
+        inverted_phase_paths.append(ref_raster)
+
+    if ts_opts.run_velocity:
+        #  We can't pass the correlations after an inversion- the numbers don't match
+        # TODO:
+        # Is there a better weighting then?
+        cor_file_list = (
+            cor_paths if len(cor_paths) == len(inverted_phase_paths) else None
+        )
+        logger.info("Estimating phase velocity")
+        timeseries.create_velocity(
+            unw_file_list=inverted_phase_paths,
+            output_file=ts_opts._velocity_file,
+            reference=reference,
+            date_list=sar_dates,
+            cor_file_list=cor_file_list,
+            cor_threshold=ts_opts.correlation_threshold,
+            num_threads=num_threads,
+        )
+
+    return inverted_phase_paths