+++ tests/test_workflows_displacement.py
@@ -25,6 +25,7 @@ def test_displacement_run_single(
             input_options={"subdataset": "/data/VV"},
             interferogram_network={
                 "indexes": [(0, -1)],
+                "max_bandwidth": 2,
             },
             phase_linking={
                 "ministack_size": 500,
@@ -42,6 +43,7 @@ def test_displacement_run_single(
         assert all(p.exists() for p in paths.stitched_cor_paths)
         assert paths.stitched_temp_coh_file.exists()
         assert paths.stitched_ps_file.exists()
+        assert paths.stitched_amp_dispersion_file.exists()
         assert paths.unwrapped_paths is not None
         assert paths.conncomp_paths is not None
         assert all(p.exists() for p in paths.conncomp_paths)